requirements.txt
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
fastapi
uvicorn
redis
chromadb
python-dotenv
# New dependencies for microservices
pyjwt
bcrypt
httpx
# Database dependencies
sqlalchemy
psycopg2-binary
alembic
# Additional LLM dependencies
openai
tiktoken
# Vector store dependencies
sentence-transformers
# Utility dependencies
python-multipart
pydantic[email]
requests
# Document processing dependencies
PyPDF2
python-docx

